[fix broken tests] use cli_runner instead of calling the method directly
